Join Project Dialog: Based it off the Create Project dialog. Made changes to coincide with what would be needed for Join Project.
Join Project Dialog.css: pulled from Join Project Dialog. Will update later but just wanted something to placehold.

NOTE: A Put method makes the most sense here to me. But api util only allows for GET or POST calls. Want to update but want claifcation if this is a necessity or if we can just POST.